pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "okta-mcp-server"
version = "0.1.0"
description = "MCP server for Okta user management"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
okta-mcp-server = "okta_mcp_server.server:main"

[tool.setuptools.packages.find]
where = ["src"]
